+++ src/js/admin.js
@@ -1,0 +1,11 @@
+import User from './user';
+
+
+class Admin extends User {
+  constructor(id, name) {
+    super(id, name);
+    this.isAdmin = true;
+  }
+}
+
+export default Admin;
