+++ src/js/eventEmitter.js
@@ -1,0 +1,24 @@
+class EventEmitter {
+  constructor() {
+    this.events = {};
+  }
+
+  subscribe(eventName, fn) {
+    if (!this.events[eventName]) {
+      this.events[eventName] = [];
+    }
+
+    this.events[eventName].push(fn);
+  }
+
+  emit(eventName, data) {
+    const event = this.events[eventName];
+    if (event) {
+      event.forEach((fn) => {
+        fn.call(null, data);
+      });
+    }
+  }
+}
+
+export default EventEmitter;
