+++ src/js/main.js
@@ -1,3 +1,3 @@
 import '../scss/main.scss';
 import '../index.html';
-import './script';
+import './index';
