+++ src/js/user.js
@@ -1,0 +1,8 @@
+class User {
+  constructor(id, name) {
+    this.id = id;
+    this.name = name;
+  }
+}
+
+export default User;
